+++ tests/test_indexes.py
@@ -1,14 +1,12 @@
 from __future__ import annotations
 
-import pandas as pd
 import pytest
-import sqlalchemy as sa
 
 from pydiverse.pipedag import Flow, Stage, Table, materialize
-from pydiverse.pipedag.context import RunContext, StageLockContext
-from pydiverse.pipedag.util.config import PipedagConfig
 
-from tests.util import select_as, tasks_library as m
+from tests.util import tasks_library as m
+
+# parameterize all tests in this file with several instance_id configurations
 from tests.fixtures.instance import *
 
 
@@ -20,6 +18,7 @@ from tests.fixtures.instance import *
         m.simple_dataframe_with_pk2,
         m.simple_dataframe_with_index,
         m.simple_dataframe_with_indexes,
+        m.simple_dataframes_with_indexes,
         m.simple_lazy_table,
         m.simple_lazy_table_with_pk,
         m.simple_lazy_table_with_pk2,