+++ CHANGELOG.rst
@@ -7,6 +7,12 @@ https://semver.org/spec/v2.0.0.html. The most important bits are:
 Changelog
 =========
 
+0.2.2 (YYYY-MM-DD)
+------------------
+
+- added option avoid_drop_create_schema to table store configuration
+- improve performance when working with IBM DB2 dialect
+
 0.2.1 (2022-01-15)
 ------------------